ktmft + refactor
